requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
pydantic[email]
python-jose
python-multipart